requirements-dev.txt
# Development dependencies
black>=21.7b0
flake8>=3.9.0
mypy>=0.910
bandit>=1.7.0
safety>=1.10.0

# Testing extras
pytest-xdist>=2.4.0
pytest-benchmark>=3.4.0
pytest-profiling>=1.7.0

# Analysis tools
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0